Updated tests to support C++20 STL

+++ test/test_string_stream_u32.cpp
@@ -68,7 +68,7 @@ namespace etl
   {
     for (auto c : str)
     {
-      os << c;
+      os << uint32_t(c);
     }
 
     return os;
